image_captioner.py
# vit-gpt2 --> trained on COCO - en cpkts
from transformers import VisionEncoderDecoderModel, ViTFeatureExtractor, AutoTokenizer
from PIL import Image
import torch


class ImageCaptioner:
    """Uses the ViT-GPT2 image captioning model from hugging face"""

    def __init__(self, max_length=30, num_beams=3):
        self.model = VisionEncoderDecoderModel.from_pretrained("nlpconnect/vit-gpt2-image-captioning")
        self.feature_extractor = ViTFeatureExtractor.from_pretrained("nlpconnect/vit-gpt2-image-captioning")
        self.tokenizer = AutoTokenizer.from_pretrained("nlpconnect/vit-gpt2-image-captioning")

        self.gen_kwargs = {"max_length": max_length, "num_beams": num_beams}
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.model.to(self.device)

    def generate_captions(self, image_paths):
        """Generate a list of captions for a given list of images"""
        images = []
        for image_path in image_paths:
            i_image = Image.open(image_path)
            if i_image.mode != "RGB":
                i_image = i_image.convert(mode="RGB")

            images.append(i_image)

        print('Generating captions...')
        pixel_values = self.feature_extractor(images=images, return_tensors="pt").pixel_values
        pixel_values = pixel_values.to(self.device)

        output_ids = self.model.generate(pixel_values, **self.gen_kwargs)

        preds = self.tokenizer.batch_decode(output_ids, skip_special_tokens=True)  # special tokens?
        preds = [pred.strip() for pred in preds]
        print('Generating captions... done')
        return preds
