requirements.txt
transformers
torch
Pillow
opencv-python